Temporary, need to fetch annotations from base
