+++ simple_parsing/annotation_utils/get_field_annotations.py
@@ -1,0 +1,185 @@
+import typing
+from typing import Dict, Any, Mapping, Union, get_type_hints
+import sys
+import collections
+from logging import getLogger as get_logger
+
+logger = get_logger(__name__)
+import builtins
+
+# NOTE: This dict is used to enable forward compatibility with things such as `tuple[int, str]`,
+# `list[float]`, etc. when using `from __future__ import annotations`.
+forward_refs_to_types = {
+    "tuple": typing.Tuple,
+    "set": typing.Set,
+    "dict": typing.Dict,
+    "list": typing.List,
+    "type": typing.Type,
+}
+import types
+import inspect
+
+
+def _replace_UnionType_with_typing_Union(annotation):
+    from simple_parsing.utils import is_list, is_tuple, is_dict, builtin_types
+
+    if sys.version_info[:2] < (3, 10):
+        # This is only useful for python 3.10+ (where UnionTypes exist).
+        # Therefore just return the annotation as-is.
+        return annotation
+
+    if isinstance(annotation, types.UnionType):  # type: ignore
+        union_args = typing.get_args(annotation)
+        new_union_args = tuple(_replace_UnionType_with_typing_Union(arg) for arg in union_args)
+        return typing.Union[new_union_args]  # type: ignore
+    if is_list(annotation):
+        item_annotation = typing.get_args(annotation)[0]
+        new_item_annotation = _replace_UnionType_with_typing_Union(item_annotation)
+        return typing.List[new_item_annotation]
+    if is_tuple(annotation):
+        item_annotations = typing.get_args(annotation)
+        new_item_annotations = tuple(
+            _replace_UnionType_with_typing_Union(arg) for arg in item_annotations
+        )
+        return typing.Tuple[new_item_annotations]  # type: ignore
+    if is_dict(annotation):
+        annotations = typing.get_args(annotation)
+        if not annotations:
+            return typing.Dict
+        assert len(annotations) == 2
+        key_annotation = annotations[0]
+        value_annotation = annotations[1]
+        new_key_annotation = _replace_UnionType_with_typing_Union(key_annotation)
+        new_value_annotation = _replace_UnionType_with_typing_Union(value_annotation)
+        return typing.Dict[new_key_annotation, new_value_annotation]
+    if annotation in builtin_types:
+        return annotation
+    if inspect.isclass(annotation):
+        return annotation
+    raise NotImplementedError(annotation)
+
+    import builtins
+
+
+#     # return forward_refs_to_types.get(ann, local_ns.get(ann, global_ns.get(ann, getattr(builtins, ann, ann))))
+
+
+def _not_supported(annotation) -> typing.NoReturn:
+    raise NotImplementedError(f"Don't yet support annotations like this: {annotation}")
+
+
+def _get_old_style_annotation(annotation: str) -> str:
+    """Replaces A | B with Union[A,B] in the annotation."""
+    # TODO: Add proper support for things like `list[int | float]`, which isn't currently
+    # working, even without the new-style union.
+    if "|" not in annotation:
+        return annotation
+
+    annotation = annotation.strip()
+    if "[" not in annotation:
+        assert "]" not in annotation
+        return "Union[" + ", ".join(v.strip() for v in annotation.split("|")) + "]"
+
+    before, lsep, rest = annotation.partition("[")
+    middle, rsep, after = rest.rpartition("]")
+    assert not after.strip(), "can't have text at HERE in <something>[<something>]<HERE>!"
+
+    if "|" in before or "|" in after:
+        _not_supported()
+    assert "|" in middle
+
+    if "," in middle:
+        parts = [v.strip() for v in middle.split(",")]
+        parts = [_get_old_style_annotation(part) for part in parts]
+        middle = ", ".join(parts)
+
+    new_middle = _get_old_style_annotation(annotation=middle)
+    new_annotation = before + lsep + new_middle + rsep + after
+    return new_annotation
+
+
+def _replace_new_union_syntax_with_old_union_syntax(
+    annotations_dict: Dict[str, str], context: collections.ChainMap
+) -> Dict[str, Any]:
+
+    new_annotations = annotations_dict.copy()
+    for field, annotation_str in annotations_dict.items():
+        updated_annotation = _get_old_style_annotation(annotation_str)
+        new_annotations[field] = updated_annotation
+
+    return new_annotations
+
+
+def get_field_type_from_annotations(some_class: type, field_name: str) -> type:
+    """
+    If the script uses `from __future__ import annotations`, and we are in python<3.9,
+    Then we need to actually first make this forward-compatibility 'patch' so that we
+    don't run into a "`type` object is not subscriptable" error.
+
+    NOTE: If you get errors of this kind from the function below, then you might want to add an
+    entry to the `forward_refs_to_types` dict above.
+    """
+
+    # Pretty hacky: Modify the type annotations of the class (preferably a copy of the class
+    # if possible, to avoid modifying things in-place), and replace  the `a | b`-type
+    # expressions with `Union[a, b]`, so that `get_type_hints` doesn't raise an error.
+    # The type of the field might be a string when using `from __future__ import annotations`.
+
+    # The type of the field might be a string when using `from __future__ import annotations`.
+    # Get the local and global namespaces to pass to the `get_type_hints` function.
+    local_ns: Dict[str, Any] = {"typing": typing, **vars(typing)}
+    if sys.version_info < (3, 9):
+        local_ns.update(forward_refs_to_types)
+    # Get the globals in the module where the class was defined.
+    global_ns = sys.modules[some_class.__module__].__dict__
+
+    try:
+        annotations_dict = get_type_hints(some_class, localns=local_ns, globalns=global_ns)
+    except TypeError as err:
+        annotations_dict = collections.ChainMap(
+            *[getattr(cls, "__annotations__", {}).copy() for cls in some_class.mro()]
+        )
+
+    if field_name not in annotations_dict:
+        raise ValueError(f"Field {field_name} not found in annotations of class {some_class}")
+
+    field_type = annotations_dict[field_name]
+
+    if sys.version_info[:2] >= (3, 7) and isinstance(field_type, typing.ForwardRef):
+        # Weird bug happens when mixing postponed evaluation of type annotations + forward
+        # references: The ForwardRefs are left as-is, and not evaluated!
+        forward_arg = field_type.__forward_arg__
+        field_type = forward_arg
+
+    if isinstance(field_type, str):
+        if "|" in field_type:
+            field_type = _get_old_style_annotation(field_type)
+
+        # Pretty hacky:
+        # In order to use `get_type_hints`, we need to pass it a class. We can't just ask it to
+        # evaluate a single annotation. Therefore, we create a temporary class and set it's
+        # __annotation__ attribute, which is introspected by `get_type_hints`.
+
+        try:
+
+            class Temp_:
+                pass
+
+            Temp_.__annotations__ = {field_name: field_type}
+            annotations_dict = get_type_hints(Temp_, globalns=global_ns, localns=local_ns)
+            field_type = annotations_dict[field_name]
+        except:
+            logger.warning(
+                f"Unable to evaluate forward reference {field_type} for field '{field_name}'.\n"
+                f"Leaving it as-is."
+            )
+            field_type = field_type
+
+    if sys.version_info >= (3, 10) and isinstance(field_type, types.UnionType):
+        # In python >= 3.10, int | float is allowed. Therefore, just to be consistent, we want
+        # to convert those into the corresponding typing.Union type.
+        # This is necessary for the rest of the code to work, since it's all based on typing.Union.
+        field_type = _replace_UnionType_with_typing_Union(field_type)
+
+    return field_type
+
